+++ lib_xcore_math/src/arch/ref/vect_inverse.c
@@ -9,10 +9,6 @@
 #include "xmath/xs3/vpu_scalar_ops.h"
 
 
-
-
-
-
 void vect_s16_inverse(
     int16_t a[],
     const int16_t b[],
@@ -20,14 +16,10 @@ void vect_s16_inverse(
     const unsigned scale)
 {
     const int32_t dividend = 1 << scale;
-    for(size_t k = 0; k < length; k++){
+    for(unsigned k = 0; k < length; k++){
         a[k] = (dividend / b[k]);
     }
 }
-
-
-
-
 
 headroom_t vect_s32_inverse(
     int32_t a[],
@@ -38,7 +30,7 @@ headroom_t vect_s32_inverse(
 
     const int64_t d = (0x1LL << scale);
 
-    for(size_t k = 0; k < length; k++){
+    for(unsigned k = 0; k < length; k++){
         a[k] = d / b[k];
         
         // printf("0x%016llX / %ld = %ld\n", (uint64_t) d, b[k], a[k]);
