+++ lib_xcore_math/src/arch/ref/vect_shl.c
@@ -9,14 +9,13 @@
 #include "xmath/xs3/vpu_scalar_ops.h"
 
 
-
 headroom_t vect_s16_shl(
     int16_t a[],
     const int16_t b[],
     const unsigned length,
     const int shl)
 {
-    for(size_t i = 0; i < length; i++){
+    for(unsigned i = 0; i < length; i++){
         a[i] = vlashr16(b[i], -shl);
     }
     return vect_s16_headroom(a, length);
@@ -31,7 +30,7 @@ headroom_t vect_s32_shl(
     const unsigned length,
     const int shl)
 {
-    for(size_t i = 0; i < length; i++){
+    for(unsigned i = 0; i < length; i++){
         a[i] = vlashr32(b[i], -shl);
     }
     return vect_s32_headroom(a, length);
